changes to Station entity to include distance to user in constructor; further implementation of ClosestStops use case

+++ src/entity/Station.java
@@ -8,7 +8,7 @@ public interface Station {
     double[] getCoordinates();
     boolean getAccessibility();
     Integer[] getTransitModes();
-    double getDistanceFromUser();
+    Integer getDistanceFromUser();
     Route[] getLines();
     ArrayList<Delay> getDelays();
     void setDelays(ArrayList<Delay> delays);
